requirements.txt
ansicon==1.89.0
arrow==1.3.0
asgiref==3.8.1
blessed==1.21.0
certifi==2025.4.26
charset-normalizer==3.4.1
Django==4.2.20
django-cron==0.6.0
django-filter==25.1
django-picklefield==3.3
django-q==1.3.9
djangorestframework==3.16.0
idna==3.10
jinxed==1.3.0
mysqlclient==2.2.7
python-dateutil==2.9.0.post0
redis==3.5.3
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
types-python-dateutil==2.9.0.20241206
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
